SecondIdmResourceTest test is removed
